% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{bayesGAMfit-class}
\alias{bayesGAMfit-class}
\alias{bayesGAMfit}
\alias{show,bayesGAMfit-method}
\title{Contains results from \code{rstan} as well as the design matrices and other data for the model.}
\usage{
\S4method{show}{bayesGAMfit}(object)
}
\arguments{
\item{object}{Object of type \code{bayesGAMfit} which stores the results from \code{rstan}, design matrices, and other data for the model.}
}
\description{
Returns object generated from model fit by \code{bayesGAM}
}
\section{Slots}{

\describe{
\item{\code{results}}{Object of type \code{stanfit} returned from calling \code{rstan::sampling}}

\item{\code{model}}{Object of custom type \code{glmModel} with the data and input parameters passed to \code{rstan}}

\item{\code{offset}}{Optionally numeric offset for the generalized additive model}

\item{\code{spcontrol}}{List of control parameters for \code{bayesGAMfit}}

\item{\code{mcmcres}}{Matrix of MCMC results for all chains, if plot data is stored}

\item{\code{pdata}}{Dataframe for default plot method, if plot data is stored.}
}}

