% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bloodpressure}
\alias{bloodpressure}
\title{Blood pressure data from a clinical study}
\format{
A data frame with 2438 rows and 13 variables:
\describe{
\item{ID}{Subject identification number}
\item{BIRTH_WT}{birth weight (lbs)}
\item{WEIGHT}{current weight (lbs)}
\item{HEIGHT}{current height (cm)}
\item{BMI}{current body mass index}
\item{age}{current age (yrs)}
\item{dias}{diastolic blood pressure}
\item{sys}{systolic blood pressure}
\item{SexM}{indicator of sex male}
\item{RaceB}{indicator of race black}
\item{RaceW}{indicator of race white}
\item{PHIGHBP}{indicator that either parent had high blood pressure}
\item{PDIABET}{indicator that either parent had diabetes}
}
}
\source{
Data provided by Wanzhu Tu, Indiana University School of Medicine
}
\usage{
bloodpressure
}
\description{
Data from 200 subjects
}
\references{
Tu, W., Eckert, G. J., DiMeglio, L. A., Yu, Z., Jung, J., and Pratt, J. H. (2011). \emph{Intensified effect of adiposity on blood pressure in overweight and obese children}. Hypertension, 58(5), 818-824.
}
\keyword{datasets}
