% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgamHMC-package.R
\docType{package}
\name{mvgamHMC-package}
\alias{mvgamHMC-package}
\title{The 'mvgamHMC' package.}
\description{
Fit Bayesian multivariate generalized additive models using Stan
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.1. https://mc-stan.org
}
