% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{reef}
\alias{reef}
\title{Coral reef data from survey data on 6 sites}
\format{
A data frame with 269 rows and 14 variables:
\describe{
\item{ZONE}{Management zone}
\item{site}{Name of the habitat site}
\item{complexity}{habitat benthic complexity}
\item{rugosity}{a measurement related to terrain complexity}
\item{LC}{cover of low complexity}
\item{HC}{cover of high complexity}
\item{SCORE1}{PCA score 1 from Wilson, Graham, Polunin}
\item{SCORE2}{PCA score 2 from Wilson, Graham, Polunin}
\item{macro}{indicator of race white}
\item{species}{fish species}
\item{abundance}{fish abundance}
\item{biomass}{fish biomass}
}
}
\source{
Data from supplementary material provided for Fisher, R., Wilson, S. K., Sin, T. M., Lee, A. C., and Langlois, T. J. (2018). \emph{A simple function for full-subsets multiple regression in ecology with R}. Ecology and evolution, 8(12), 6104-6113.
}
\usage{
reef
}
\description{
Data from 68 subjects
}
\references{
Wilson, S. K., Graham, N. A. J., and Polunin, N. V. (2007). \emph{Appraisal of visual assessments of habitat complexity and benthic composition on coral reefs}. Marine Biology, 151(3), 1069-1076.
}
\keyword{datasets}
