% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{st}
\alias{st}
\title{Constructor function for Student-t priors}
\usage{
st(param_values)
}
\arguments{
\item{param_values}{Numeric vector of length 3 for the degrees of freedom, location, and scale parameter.}
}
\description{
Used to specify student-t priors for \code{bayesGAM} models
}
\details{
For the \code{beta} and \code{a} parameters, the distribution is assumed to be unconstrained.
For \code{eps} and \code{lambda}, the priors are half-normal with a support of strictly positive numbers.
}
\examples{
require(stats); require(graphics)
st(c(3,0,1))
}
\references{
Stan Development Team. 2018. Stan Modeling Language Users Guide and Reference Manual, Version 2.18.0
}
