% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{summary}
\alias{summary}
\alias{summary,mvgamHMCfit-method}
\title{Summarizing Model Fits from \code{mvgamHMC}}
\usage{
\S4method{summary}{mvgamHMCfit}(object)
}
\arguments{
\item{object}{an object of class \code{hmclearn}, usually a result of a call to \code{mh} or \code{hmc}}
}
\value{
Returns a matrix with posterior quantiles and the posterior scale reduction factor statistic for each parameter.
}
\description{
summary method for class \code{mvgamHMCfit}
}
\examples{
f <- mvgamHMC(weight ~ np(height), data = women, 
              family = gaussian, iter=500, chains = 1)

summary(f)
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.1.
}
