% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{getStanResults}
\alias{getStanResults}
\alias{getStanResults,mvgamHMCfit-method}
\title{Returns the \code{stanfit} object generated by \pkg{rstan}}
\usage{
getStanResults(object)

\S4method{getStanResults}{mvgamHMCfit}(object)
}
\arguments{
\item{object}{Object of type \code{mvgamHMCfit} returned from \code{mvgamHMC}}
}
\value{
Contents of \code{stanfit} results
}
\description{
Contains the full content of the \code{stanfit} object
}
\examples{
require(stats); require(graphics)
f <- mvgamHMC(weight ~ np(height), data = women, family = gaussian, 
              iter = 500, chains = 1)
sres <- getStanResults(f)
plot(sres) # rstan method
}
